requirements.txt
streamlit
openai
python-dotenv
pillow
requests
beautifulsoup4
pyPDF2
docx
